rename: features>ktWiz 폴더 이름 변경(#191)
