+++ src/pages/ktwiz/KtHistoryPage.tsx
@@ -1,21 +1,21 @@
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui";
-import { useTabFromUrl } from "@/hooks/useTabFromUrl";
-import { cn } from "@/lib/utils";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui';
+import { useTabFromUrl } from '@/hooks/useTabFromUrl';
+import { cn } from '@/lib/utils';
 
-import Banner from "@/features/common/Banner";
-import Layout from "@/features/common/Layout";
-import { TeamHistoryTab, TeamIntroduceTab } from "@/features/kwWiz";
+import Banner from '@/features/common/Banner';
+import Layout from '@/features/common/Layout';
+import { TeamHistoryTab, TeamIntroduceTab } from '@/features/ktWiz';
 
 const WIZ_TABS_CONFIG = [
-  { value: "about", path: "/about" },
-  { value: "history", path: "/history" },
+  { value: 'about', path: '/about' },
+  { value: 'history', path: '/history' },
 ];
 
 const KtHistoryPage = () => {
   const { currentTab, handleTabChange } = useTabFromUrl({
-    basePath: "/ktwiz",
+    basePath: '/ktwiz',
     tabs: WIZ_TABS_CONFIG,
-    defaultTab: "history",
+    defaultTab: 'history',
   });
 
   return (
@@ -41,19 +41,19 @@ const KtHistoryPage = () => {
         defaultValue={currentTab}
         onValueChange={handleTabChange}
       >
-        <div className={cn("header")}>
+        <div className={cn('header')}>
           {/* 탭 */}
           <div className="tabs-wrapper">
             <TabsList className="tabs-list">
               <TabsTrigger
                 value="about"
-                onClick={() => handleTabChange("about")}
+                onClick={() => handleTabChange('about')}
               >
                 구단 소개
               </TabsTrigger>
               <TabsTrigger
                 value="history"
-                onClick={() => handleTabChange("history")}
+                onClick={() => handleTabChange('history')}
               >
                 구단 연혁
               </TabsTrigger>
